requirements.txt
openai>=1.0.0
feedparser>=6.0.0
httpx>=0.27.0
firecrawl-py>=1.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
